Added app diagrams, Refactored class names, added comments
