+++ java-samples/corda5-negotiation-cordapp/contracts/src/main/java/com/r3/developers/samples/negotiation/states/Trade.java
@@ -1,9 +1,8 @@
-package com.r3.developers.samples.obligation.states;
+package com.r3.developers.samples.negotiation.states;
 
 
-import com.r3.developers.samples.obligation.contracts.ProposalAndTradeContract;
+import com.r3.developers.samples.negotiation.contracts.ProposalAndTradeContract;
 import net.corda.v5.base.annotations.ConstructorForDeserialization;
-import net.corda.v5.base.types.MemberX500Name;
 import net.corda.v5.ledger.utxo.BelongsToContract;
 import net.corda.v5.ledger.utxo.ContractState;
 import org.jetbrains.annotations.NotNull;
@@ -19,16 +18,16 @@ public class Trade implements ContractState {
     private int amount;
     private Member buyer;
     private Member seller;
-    private UUID linearId;
+    private UUID proposalID;
     public List<PublicKey> participants;
 
 
     @ConstructorForDeserialization
-    public Trade(int amount, Member buyer, Member seller, UUID linearId, List<PublicKey> participants) {
+    public Trade(int amount, Member buyer, Member seller, UUID proposalID, List<PublicKey> participants) {
         this.amount = amount;
         this.buyer = buyer;
         this.seller = seller;
-        this.linearId = linearId;
+        this.proposalID = proposalID;
         this.participants = participants;
     }
 
@@ -36,7 +35,7 @@ public class Trade implements ContractState {
         this.amount = amount;
         this.buyer = buyer;
         this.seller = seller;
-        this.linearId = UUID.randomUUID();
+        this.proposalID = UUID.randomUUID();
         this.participants = participants;
     }
 
@@ -51,8 +50,8 @@ public class Trade implements ContractState {
         return buyer;
     }
 
-    public UUID getLinearId() {
-        return linearId;
+    public UUID getProposalID() {
+        return proposalID;
     }
 
     @NotNull
