+++ java-samples/corda5-negotiation-cordapp/workflows/src/main/java/com/r3/developers/samples/negotiation/workflows/accept/AcceptFlowArgs.java
@@ -1,4 +1,4 @@
-package com.r3.developers.samples.obligation.workflows.accept;
+package com.r3.developers.samples.negotiation.workflows.accept;
 
 import net.corda.v5.base.annotations.CordaSerializable;
